% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abfall_zh}
\alias{abfall_zh}
\title{Abfallmengen nach Abfallarten und Gemeinden Kanton Zuerich}
\format{
Ein Datensatz mit 6 Variablen und 34668 Zeilen:
\describe{
\item{Jahr}{Jahr der Erhebung}
\item{Gemeinde_Nr}{Kantonale Gemeinde ID}
\item{Gemeinde}{Name der Gemeinde}
\item{Abfallart}{Spezifizierung des Abfalles}
\item{Wert}{numerische Mengenangabe}
\item{Einheit}{Masseinheit des Wertes}
}
}
\usage{
abfall_zh
}
\description{
Source: https://www.zh.ch/de/politik-staat/opendata.zhweb-noredirect.zhweb-cache.html?keywords=ogd#/datasets/1803@awel-kanton-zuerich/distributions/3460
}
\keyword{datasets}
