% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_amounts.R
\name{plot_amounts}
\alias{plot_amounts}
\alias{plot_amounts_horizontal}
\alias{plot_amounts_grouped}
\alias{plot_amounts_facets}
\alias{plot_amounts_stacked}
\title{Mengen visualisieren}
\usage{
plot_amounts(data, x, y, color = "#0d7abc")

plot_amounts_horizontal(data, x, y, color = "#0d7abc")

plot_amounts_grouped(data, x, y, group)

plot_amounts_facets(data, x, y, facet, n_col = 2, color = "#0d7abc")

plot_amounts_stacked(data, x, y, group)
}
\arguments{
\item{data}{Ein Tibble mit den Daten f端r den Plot.}

\item{x}{Die Variable f端r die x-Achse.}

\item{y}{Die Variable f端r die y-Achse.}

\item{color}{Farbe der Balken. Default = #0d7abc.}

\item{group}{Gruppierungsvariable f端r einen einzelnen Plot.}

\item{facet}{Gruppierungsvariable um mehrere Plots zu machen (Facetten).}

\item{n_col}{Anzahl Spalten bei den Facetten. Default = 2.}
}
\value{
ggplot object
}
\description{
Mengen visualisieren
}
\examples{
# Test Daten als Beispiel verwenden
df <-
tibble::tibble(
 x = c("a", "b", "c", "d", "e"),
 y = 1:5
)

plot_amounts(df, x, y)

}
