% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distributions.R
\name{plot_distributions}
\alias{plot_distributions}
\alias{plot_distributions_density}
\alias{plot_distributions_sidebyside}
\alias{plot_distributions_grouped}
\alias{plot_distributions_raincloud}
\alias{plot_distributions_boxplot}
\title{Verteilungen visualisieren}
\usage{
plot_distributions(data, variable, bw = 5, color = "#0d7abc")

plot_distributions_density(data, variable, bw = 5, color = "#0d7abc")

plot_distributions_sidebyside(data, variable, group, bw = 5)

plot_distributions_grouped(data, variable, group, bw = 5)

plot_distributions_raincloud(data, variable, group, bw = 5, color = "#0d7abc")

plot_distributions_boxplot(
  data,
  variable,
  group,
  bw = 5,
  color = "#0d7abc",
  size = 5
)
}
\arguments{
\item{data}{Ein Tibble mit den Daten für den Plot.}

\item{variable}{Variable die untersucht wird (x-Achse).}

\item{bw}{Breite (Anzahl werte die zusammengefasst werden) der Behälter.}

\item{color}{Farbe um die Balken optisch zu trennen.}

\item{group}{Variable um die Daten zu gruppieren.}

\item{size}{Grösse der Schrift.}
}
\value{
ggplot2 object
}
\description{
Verteilungen visualisieren
}
\examples{
#tbd

}
