% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proportions.R
\name{plot_proportions}
\alias{plot_proportions}
\alias{plot_proportions_donut}
\alias{plot_proportions_stacked}
\alias{plot_proportions_sidebyside_bar}
\alias{plot_proportions_sidebyside_density1}
\alias{plot_proportions_sidebyside_density2}
\title{Proportionen visualisieren}
\usage{
plot_proportions(data, n = n, group)

plot_proportions_donut(data, n = n, percent = percent, group)

plot_proportions_stacked(data, x, y, group)

plot_proportions_sidebyside_bar(data, x, percent, facet)

plot_proportions_sidebyside_density1(
  data,
  x,
  group,
  bw = 5,
  color = "#0d7abc",
  facet_row = 1
)

plot_proportions_sidebyside_density2(
  data,
  x,
  group,
  bw = 5,
  color = "#0d7abc",
  facet_row = 1
)
}
\arguments{
\item{data}{Ein Tibble mit den Daten für den Plot.}

\item{n}{Der Wert zum ploten.}

\item{group}{Gruppierungsvariable für einen einzelnen Plot.}

\item{percent}{Prozentangabe als Dezimalzahl (0.0 bis 1.0)}

\item{x}{Die Variable für die x-Achse.}

\item{y}{Die Variable für die y-Achse.}

\item{facet}{Gruppierungsvariable um einen Plot je Teilaspekt zu machen (Facetten).}

\item{bw}{Breite (Anzahl werte die zusammengefasst werden) der Behälter.}

\item{color}{Farbe um die Balken optisch zu trennen.}

\item{facet_row}{Anzahl Zeilen für Darstellung der Teilaspekte.}
}
\value{
ggplot object
}
\description{
Proportionen visualisieren
}
\examples{
# tbd

}
