% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_series.R
\name{plot_timeseries_slope}
\alias{plot_timeseries_slope}
\alias{plot_timeseries_line}
\alias{plot_timeseries_trend}
\alias{plot_timeseries_detrend}
\title{Zeitreihen visualisieren. plot_timeseries_* Familie.}
\usage{
plot_timeseries_slope(
  data,
  x,
  y,
  group,
  color_point = "#0d7abc",
  color_line = "#b3b3b3a0"
)

plot_timeseries_line(
  data,
  x,
  y,
  group,
  color_line = "#0d7abc",
  color_area = "#6fa8dc",
  y_limits = NULL
)

plot_timeseries_trend(
  data,
  x,
  y,
  method = NULL,
  formula = NULL,
  color_line = "black",
  color_trend = "#0d7abc",
  y_limits = NULL
)

plot_timeseries_detrend(
  data,
  x,
  y,
  method = "pc",
  group,
  facet = TRUE,
  color_line = "#0d7abc",
  color_area = "#6fa8dc",
  y_limits = NULL
)
}
\arguments{
\item{data}{Ein Tibble mit den Daten für den Plot.}

\item{x}{Die Variable für die x-Achse.}

\item{y}{Die Variable für die y-Achse.}

\item{group}{Gruppierungsvariable für einen einzelnen Plot.}

\item{color_point}{Farbe der Punkte.}

\item{color_line}{Farbe für die Linie.}

\item{color_area}{Farbe vom Bereich unter dem Linienplot.}

\item{y_limits}{Die beiden Poole bei der Y-Achse.
\href{https://ggplot2.tidyverse.org/reference/scale_continuous.html?q=scale_y_con#arguments}{Mehr Infos im Kapitel Arguments beim Argument limits in der ggplot2 Dokumentation}}

\item{method}{Methode die in der Funktion verwendet wird. Variiert je nach Funktion.
\href{https://ggplot2.tidyverse.org/reference/geom_smooth.html?q=geom_smooth#arguments}{Mehr Infos im Kapitel Arguments beim Argument method in der ggplot2 Dokumentation}}

\item{formula}{Berechnungsformel der Methode.
\href{https://ggplot2.tidyverse.org/reference/geom_smooth.html?q=geom_smooth#arguments}{Mehr Infos im Kapitel Arguments beim Argument formula in der ggplot2 Dokumentation}}

\item{color_trend}{Farbe für die Trendlinie.}

\item{facet}{Gruppierungsvariable um einen Plot je Teilaspekt zu machen (Facetten).}
}
\value{
ggplot object
}
\description{
Zeitreihen visualisieren. plot_timeseries_* Familie.
}
\examples{
ggplot2::mpg |>
  dplyr::group_by(year, manufacturer) |>
  dplyr::summarise(hwy_by_manufacturer = mean(hwy)) |>
  dplyr::ungroup() |>
  plot_timeseries_slope(x = year, y = hwy_by_manufacturer, group = manufacturer)

stock_markets <- tsbox::ts_tbl(datasets::EuStockMarkets) |>
  dplyr::mutate(time = lubridate::as_date(time))

# ein Aktienmarkt (1 Merkmal)
stock_markets |>
  dplyr::filter(id == "FTSE") |>
  plot_timeseries_line(x = time, y = value, group = id)

# mehrere Aktienmaerkte (n Merkmale)
stock_markets |>
  dplyr::filter(id != "FTSE") |>
  plot_timeseries_line(x = time, y = value, group = id)

# ein Aktienmarkt (1 Merkmal)
stock_markets |>
  dplyr::filter(id == "FTSE") |>
  plot_timeseries_trend(x = time, y = value)

# mehrere Aktienmaerkte (n Merkmale)
stock_markets |>
  dplyr::filter(id != "FTSE") |>
  plot_timeseries_trend(x = time, y = value) +
  # Erweiterung mit ggplot2 Funktion
  ggplot2::facet_wrap(ggplot2::vars(id))

# trendbereinigte monatliche Entwicklung
stock_markets |>
  # zuerst die durchschnittliche
  # monatliche entwicklung berechnen
  dplyr::mutate(time = format(time, "\%Y-\%m")) |>
  dplyr::group_by(id, time) |>
  dplyr::summarise(value = mean(value)) |>
  dplyr::ungroup() |>
  plot_timeseries_detrend(x = time, y = value, group = id)

}
