% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distributions.R
\name{plot_distributions_histogram}
\alias{plot_distributions_histogram}
\alias{plot_distributions_density}
\alias{plot_distributions_sidebyside}
\alias{plot_distributions_grouped}
\alias{plot_distributions_raincloud}
\alias{plot_distributions_boxplot}
\title{Verteilungen visualisieren. plot_distributions_* Familie.}
\usage{
plot_distributions_histogram(data, x, bw = 5, color = "#0d7abc")

plot_distributions_density(data, x, bw = 5, color = "#0d7abc")

plot_distributions_sidebyside(data, x, group, bw = 5)

plot_distributions_grouped(data, x, group, bw = 5)

plot_distributions_raincloud(data, x, y, bw = 5, color = "#0d7abc")

plot_distributions_boxplot(data, x, y, bw = 5, color = "#0d7abc", size = 5)
}
\arguments{
\item{data}{Ein Tibble mit den Daten für den Plot.}

\item{x}{Variable die untersucht wird (x-Achse).}

\item{bw}{Breite (Anzahl werte die zusammengefasst werden) der Behälter.
\href{https://ggplot2.tidyverse.org/reference/geom_density.html?q=density#arguments}{Mehr Infos beim Argument bw in der ggplot2 Dokumentation}}

\item{color}{Farbe um die Balken optisch zu trennen.}

\item{group}{Variable um die Daten zu gruppieren.}

\item{y}{Variable die untersucht wird (y-Achse).}

\item{size}{Grösse der Schrift.}
}
\value{
ggplot2 object
}
\description{
Verteilungen visualisieren. plot_distributions_* Familie.
}
\examples{
df <- socviz::gss_lon

cowplot::plot_grid(
  plot_distributions_histogram(data = df, x = age),
  plot_distributions_histogram(data = df, x = age, bw = 2),
  labels = c('A: bw = 5 (default)', 'B: bw = 2')
  )

# mit pipe syntax
df |>
  tidyr::drop_na(happy) |>
  plot_distributions_grouped(age, group = happy, bw = 3)

# ohne pipe
  plot_distributions_sidebyside(data = df, x = age, group = sex)

df |>
  # jede 100th Zeile im df behalten
  # notwendig, da ansonsten zu grosse
  # datenmenge für diesen plot
  dplyr::filter(dplyr::row_number() \%\% 100 == 1) |>
  plot_distributions_raincloud(x = happy, y = age)

cowplot::plot_grid(
  plot_distributions_raincloud(data = df, x = happy, y = age),
  plot_distributions_boxplot(data = df, x = happy, y = age, size = 3),
  labels = c("A: Zuviele Datenpunkte für raincloud plot.",
             "B: Beobachtungen als Zahl darstellen."
             )
  )


}
