% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proportions.R
\name{plot_proportions_bar}
\alias{plot_proportions_bar}
\alias{plot_proportions_donut}
\alias{plot_proportions_stacked}
\alias{plot_proportions_sidebyside_bar}
\alias{plot_proportions_sidebyside_density1}
\alias{plot_proportions_sidebyside_density2}
\title{Proportionen visualisieren. plot_proportions_* Familie.}
\usage{
plot_proportions_bar(data, group, percent = percent, n = n)

plot_proportions_donut(data, group, n = n, percent = percent)

plot_proportions_stacked(data, x, y, group)

plot_proportions_sidebyside_bar(data, x, y, color, facet)

plot_proportions_sidebyside_density1(
  data,
  x,
  group,
  bw = 5,
  color = "#0d7abc",
  facet_row = 1
)

plot_proportions_sidebyside_density2(
  data,
  x,
  group,
  bw = 5,
  color = "#0d7abc",
  facet_row = 1
)
}
\arguments{
\item{data}{Ein Tibble mit den Daten für den Plot.}

\item{group}{Gruppierungsvariable für einen einzelnen Plot.}

\item{percent}{Prozentangabe als Dezimalzahl (0.0 bis 1.0)}

\item{n}{Der Wert zum ploten.}

\item{x}{Die Variable für die x-Achse.}

\item{y}{Die Variable für die y-Achse.}

\item{color}{Farbe um Gruppierungen zu markieren.}

\item{facet}{Gruppierungsvariable um einen Plot je Teilaspekt zu machen (Facetten).}

\item{bw}{Breite (Anzahl werte die zusammengefasst werden) der Behälter.
\href{https://ggplot2.tidyverse.org/reference/geom_density.html?q=density#arguments}{Mehr Infos beim Argument bw in der ggplot2 Dokumentation}}

\item{facet_row}{Anzahl Zeilen für Darstellung der Teilaspekte.}
}
\value{
ggplot object
}
\description{
"Hinweis: Daten in den Dichteplots (density) duerfen nicht aggregiert sein.
Die Berechnung erfolgt in der Funktion.
}
\examples{
df1 <- tibble::tibble(
  x = c("a", "b", "c"),
  y = c(3, 7, 10)
  )

df1 |>
  dplyr::mutate(percent = y / sum(y)) |>
  plot_proportions_bar(group = x, n = y, percent = percent)

# aufbereitung der daten
df2 <- socviz::gss_sib |>
  dplyr::count(year, sex) |>
  dplyr::group_by(year) |>
  dplyr::mutate(sum_year = sum(n)) |>
  dplyr::group_by(year, sex) |>
  dplyr::summarise(percent = n / sum_year) |>
  dplyr::ungroup()

# mit pipe syntax
df2 |>
  plot_proportions_stacked(x = year, y = percent, group = sex)

# ohne pipe
plot_proportions_sidebyside_bar(data = df2, x = year, y = percent, color = sex, facet = sex)

socviz::gss_lon |>
  plot_proportions_sidebyside_density1(x = as.numeric(age), group = happy)


}
