% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_series.R
\name{plot_timeseries_slope}
\alias{plot_timeseries_slope}
\alias{plot_timeseries_line}
\alias{plot_timeseries_trend}
\alias{plot_timeseries_detrend}
\title{Zeitreihen visualisieren}
\usage{
plot_timeseries_slope(
  data,
  x,
  y,
  group,
  color_point = "#0d7abc",
  color_line = "#b3b3b3a0"
)

plot_timeseries_line(
  data,
  x,
  y,
  group,
  color_line = "#0d7abc",
  color_area = "#6fa8dc",
  y_limits = NULL
)

plot_timeseries_trend(
  data,
  x,
  y,
  method = NULL,
  formula = NULL,
  span = 0.75,
  color_line = "black",
  color_trend = "#0d7abc",
  y_limits = NULL
)

plot_timeseries_detrend(
  data,
  x,
  y,
  method = "pc",
  group,
  facet = TRUE,
  color_line = "#0d7abc",
  color_area = "#6fa8dc",
  y_limits = NULL
)
}
\arguments{
\item{data}{Ein Tibble mit den Daten für den Plot.}

\item{x}{Die Variable für die x-Achse.}

\item{y}{Die Variable für die y-Achse.}

\item{group}{Gruppierungsvariable für einen einzelnen Plot.}

\item{color_point}{Farbe der Punkte.}

\item{color_line}{Farbe für die Linie.}

\item{color_area}{Farbe vom Bereich unter dem Linienplot.}

\item{y_limits}{Die beiden Poole bei der Y-Achse.}

\item{method}{Methode die in der Funktion verwendet wird. Variiert je nach Funktion.}

\item{formula}{Berechnungsformel der Methode.}

\item{span}{Zeitspanne einer Analyseeinheit.}

\item{color_trend}{Farbe für die Trendlinie.}

\item{facet}{Gruppierungsvariable um einen Plot je Teilaspekt zu machen (Facetten).}
}
\value{
ggplot object
}
\description{
Zeitreihen visualisieren
}
\examples{
# tbd

}
